\name{Calomys}
\alias{Calomys}
\docType{data}
\title{Calomys phylogeny}
\description{
Ultrametric phylogenetic tree of 11 of the 13 extant Calomys species
}
\usage{data(Calomys)}
\details{This phylogeny is from Pigot et al. PloS Biol 2012}

\references{
Pigot et al.(2012) Speciation and extinction drive the appearance of directional range size evolution in phylogenies and the fossil record \emph{PloS Biol} 10:1-9

Manceau, M., Lambert, A., Morlon, H. (submitted)
}

\examples{
data(Calomys)
print(Calomys)
plot(Calomys)
}
\keyword{datasets}
