\name{Cetacea}
\alias{Cetacea}
\docType{data}
\title{Cetacean phylogeny}
\description{Ultrametric phylogenetic tree for 87 of the 89 extant cetacean species}
\usage{data(Cetacea)}
\details{
This phylogeny was constructed by Bayesian phylogenetic inference from six mitochondrial and nine nuclear genes.
It was calibrated using seven paleontological age constraints and a relaxed molecular clock approach. See Steeman et al. (2009) for details.
}

\source{
Steeman ME et al.(2009) Radiation of extant cetaceans driven by restructuring of the oceans, \emph{Syst Biol} 58:573-585
}

\references{
Steeman ME et al.(2009) Radiation of extant cetaceans driven by restructuring of the oceans \emph{Syst Biol} 58:573-585

Morlon, H., Parsons, T.L., Plotkin, J.B. (2011) Reconciling molecular phylogenies with the fossil record \emph{Proc Nat Acad Sci} 108: 16327-16332

Condamine, F.L., Rolland, J., Morlon, H. (2013) Macroevolutionary perspectives to environmental change \emph{Eco Lett} 16: 72-85
}
\examples{
data(Cetacea)
print(Cetacea)
plot(Cetacea)
}
\keyword{datasets}
