\name{Phocoenidae}
\alias{Phocoenidae}
\docType{data}
\title{Phocoenidae phylogeny}
\description{
Ultrametric phylogenetic tree of the 6 extant Phocoenidae (porpoise) species
}
\usage{data(Phocoenidae)}
\details{This phylogeny was extracted from Steeman et al. Syst Bio 2009 cetacean phylogeny}

\references{
Steeman ME et al.(2009) Radiation of extant cetaceans driven by restructuring of the oceans \emph{Syst Biol} 58:573-585

Morlon, H., Parsons, T.L., Plotkin, J.B. (2011) Reconciling molecular phylogenies with the fossil record \emph{Proc Nat Acad Sci} 108: 16327-16332
}

\examples{
data(Phocoenidae)
print(Phocoenidae)
plot(Phocoenidae)
}
\keyword{datasets}
