\name{Phyllostomidae}
\alias{Phyllostomidae}
\docType{data}
\title{Phyllostomidae phylogeny}
\description{
Ultrametric phylogenetic tree of 150 of the 165 extant known Phyllostomidae species
}
\usage{data(Phyllostomidae)}
\details{This phylogeny is the maximum clade credibility tree used in Rolland et al. (2014), which originally comes from the Bininda-Emonds tree (Bininda-Emonds et al. 2007)}

\references{
Bininda-Emonds, O. R., et al. (2007) The delayed rise of present-day mammals \emph{Nature} 446: 507-512

Rolland, J., Condamine, F. L., Jiguet, F., & Morlon, H. (2014) Faster speciation and reduced extinction in the tropics contribute to the mammalian latitudinal diversity gradient. \emph{PLoS Biol}, 12(1): e1001775.
}

\seealso{
\code{\link{Phyllostomidae_genera}}}

\examples{
data(Phyllostomidae)
print(Phyllostomidae)
#plot(Phyllostomidae)
}
\keyword{datasets}
