\name{Phyllostomidae_genera}
\alias{Phyllostomidae_genera}
\docType{data}
\title{Phylogenies of Phyllostomidae genera}
\description{
List of 25 ultrametric phylogenetic trees corresponding to 25 Phyllostomidae genera
}
\usage{data(Phyllostomidae_genera)}

\seealso{
\code{\link{Phyllostomidae}}}

\examples{
data(Phyllostomidae_genera)
print(Phyllostomidae_genera)
}
\keyword{datasets}
